+++ packages/cooking/util/load-postcss.js
@@ -5,5 +5,7 @@ const isArray = require('./is').Array
 module.exports = function (plugins) {
   return isArray(plugins) ?
     () => plugins :
-    () => plugins.apply(this, arguments)
+    function () {
+      return plugins.apply(this, arguments)
+    }
 }